+++ src/collective/cover/content.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-# XXX: remove these imports?
-#from plone.dexterity.events import EditCancelledEvent
-#from plone.dexterity.events import EditFinishedEvent
 from AccessControl import getSecurityManager
 from Acquisition import aq_inner
 from collective.cover.controlpanel import ICoverSettings
+from collective.cover.interfaces import IGridSystem
 from collective.cover.utils import assign_tile_ids
 from five import grok
 from plone.dexterity.content import Item
@@ -201,6 +199,13 @@ class LayoutEdit(grok.View):
         portal = getToolByName(self.context, 'portal_url').getPortalObject()
         # TODO: check permission locally and not in portal context
         return sm.checkPermission('collective.cover: Can Export Layout', portal)
+
+    def layoutmanager_settings(self):
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ICoverSettings)
+        grid = getUtility(IGridSystem, name=settings.grid_system)
+
+        return json.dumps({'ncolumns': grid.ncolumns})
 
 
 class UpdateTileContent(grok.View):