+++ src/collective/cover/interfaces.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from zope.interface import Attribute
 from zope.interface import Interface
 
 
@@ -35,3 +36,11 @@ class ICoverUIDsProvider(Interface):
 class ITileEditForm(Interface):
     """Custom EditForm interface for a tile.
     """
+
+
+class IGridSystem(Interface):
+    """Interface for classes that implement a grid system for collective
+    cover."""
+
+    title = Attribute('The user-visible title for this grid.')
+    ncolums = Attribute('Number of colums in a grid.')