+++ src/collective/cover/tests/test_vocabularies.py
@@ -2,6 +2,7 @@
 
 from collective.cover.controlpanel import ICoverSettings
 from collective.cover.testing import INTEGRATION_TESTING
+from collective.cover.testing import MULTIPLE_GRIDS_INTEGRATION_TESTING
 from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
 from zope.component import queryUtility
@@ -99,7 +100,39 @@ class VocabulariesTestCase(unittest.TestCase):
         self.assertEqual(styles.by_value.keys()[0], u'tile-default')
         # adding a couple of not well formatted items result in no option
         # (except for the default one)
-        settings.styles = set(['not well formated'])
+        settings.styles = set(['not well formatted'])
         styles = vocabulary(self.portal)
         self.assertEqual(len(styles), 1)
         self.assertEqual(styles.by_value.keys()[0], u'tile-default')
+
+    def test_grid_systems(self):
+        name = u'collective.cover.GridSystems'
+        vocabulary = queryUtility(IVocabularyFactory, name)
+        self.assertIsNotNone(vocabulary)
+
+        # Our default grid system must be in the vocabulary.
+        grids = vocabulary(self.portal)
+        self.assertEqual(len(grids), 1)
+        self.assertIn(u'deco16_grid', grids)
+
+
+class VocabulariesMultipleGridsTestCase(unittest.TestCase):
+
+    layer = MULTIPLE_GRIDS_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+
+    def test_grid_systems(self):
+        name = u'collective.cover.GridSystems'
+        vocabulary = queryUtility(IVocabularyFactory, name)
+        self.assertIsNotNone(vocabulary)
+        grids = vocabulary(self.portal)
+        self.assertEqual(len(grids), 2)
+
+        # Our default grid system must be in the vocabulary.
+        self.assertIn(u'deco16_grid', grids)
+
+        # The layer has setup a second grid.
+        self.assertIn(u'bootstrap3', grids)
+        self.assertEqual(grids.getTerm('bootstrap3').title, u'Bootstrap 3')